[refactor] common폴더명 style로 변경
- style에 대한 파일들이라 style로 폴더명 변경
