+++ src/components/UI/atom/Anchor/Anchor.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { withKnobs, text, color } from '@storybook/addon-knobs';
 import Anchor from '.';
-import StyleWrapper from '../../../../common/styleWrapper';
-import theme from '../../../../common/theme';
+import StyleWrapper from '../../../../style/styleWrapper';
+import theme from '../../../../style/theme';
 
 export default {
   title: 'Atom/Anchor',
