+++ src/components/UI/atom/Button/Button.stories.tsx
@@ -3,7 +3,7 @@ import { withKnobs, text } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import Button from '.';
 import Icon from '../Icon';
-import StyleWrapper from '../../../../common/styleWrapper';
+import StyleWrapper from '../../../../style/styleWrapper';
 
 export default {
   title: 'Atom/Button',
