+++ src/components/UI/atom/Tag/Tag.stories.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { withKnobs, text, color } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import Tag from '.';
-import StyleWrapper from '../../../../common/styleWrapper';
-import theme from '../../../../common/theme';
+import StyleWrapper from '../../../../style/styleWrapper';
+import theme from '../../../../style/theme';
 
 export default {
   title: 'Atom/Tag',
