+++ src/components/UI/atom/Tag/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as Style from './styled';
-import theme from '../../../../common/theme';
+import theme from '../../../../style/theme';
 
 export interface TagProps {
   id: string;
