+++ src/components/UI/organism/ProfileIntroduce/styled.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import theme from '../../../../common/theme';
+import theme from '../../../../style/theme';
 
 export const Form = styled.form`
   height: 12.5rem;
