+++ src/components/UI/organism/SignupForm/styled.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import theme from '../../../../common/theme';
+import theme from '../../../../style/theme';
 
 export const Container = styled.form`
   width: 320px;
