+++ src/components/UI/organism/SummarizedPostItem/styled.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import theme from '../../../../common/theme';
+import theme from '../../../../style/theme';
 
 interface FinishTagProps {
   isFinished: boolean;
