+++ src/index.tsx
@@ -5,7 +5,7 @@ import axios from 'axios';
 import { Provider } from 'react-redux';
 
 import App from './App';
-import GlobalStyle from './common/GlobalStyle';
+import GlobalStyle from './style/GlobalStyle';
 import { SERVER_URL } from './config';
 import store from './redux/store';
